2026 / 04 / 16

+++ 자바프로그래밍 (2-1)/십습/4주차/ByteShiftTest.java
@@ -1,7 +1,7 @@
 public class ByteShiftTest {
 	public static void main(String[] args) {
 		for (int i = 0; i < 32; i++) {
-			System.out.println((byte)Integer.toString(1<<i));
+			System.out.println(Integer.toString(1<<i));
 		}
 	}
 }
